Added notes

+++ Conjugate Gradients Spyder/cg_noprecon.py
@@ -20,23 +20,36 @@ def protein_cg(lengthscale_range, noise_range, kernel, ARD=False, n=1000):
     
     #if ARD then we have E[lengthscales] = lengthscale_range
     if ARD:
+        #initiate covariance function
         covar_fn = kernel(ard_num_dims=9)
+        #store lengthscales in array
         lengthscales = np.zeros((num_trials, 9))
         for i in range(num_trials):
+            
             if i < num_trials-1:
                 dist = lengthscale_range[i+1] - lengthscale_range[i]
-            lengthscales[i] = abs( np.random.normal(lengthscale_range[i], dist, [1,9]).astype('float32') )
+            
+            #compute lengthscales for current batch
+            lengthscales[i] = abs( np.random.normal(lengthscale_range[i], dist, [1,9]) )
             ave = np.average(lengthscales[i])
+            
+            #initialise lengthscales in cov
             covar_fn.lengthscale = lengthscales[i] 
             x_tens = torch.Tensor(train_x)
+            #compute cov
             K_XX = covar_fn(x_tens, x_tens).evaluate().detach().numpy()
             for j in range(noise_length):
+                #assign noise & compute likelihood cov
                 sigma = noise_range[j]
                 K_hat = K_XX + sigma ** 2 * np.eye(n)
+                #compute exact solution for error comparison
                 sol = np.linalg.inv(K_hat) @ train_y
+                #run cg
                 cg_run = cg_test(lin_cg, K_hat, train_y, sol=sol)
+                #store results
                 results = [[ave, sigma, cg_run[0], cg_run[1]]]
                 res = np.vstack([res, results])
+        #return results as a dataframe
         res = pandas.DataFrame( res[1:, 0:3], columns=['Lengthscales', 'Noise', 'Iterations'])
         return res
